Unified autodiff gradient path, muPC scaling at callsites, package restructure (#20)
## TL;DR
Internal-infrastructure release. Three threads converge: (1) node gradient computation is now pure autodiff with muPC scaling applied by the inference/learning loops, (2) the `builder` package is dissolved into `core.topology` + `graph_assembly`, breaking a long-standing circular import and tightening the dependency DAG, (3) integer clamps (e.g. token indices) flow end-to-end without silent dtype promotion. 127/127 tests pass; demos run clean.

## Why now
The original design carried muPC scaling and self-gradient accumulation inside each node's forward methods. Every new node type had to re-derive how those interacted, and `_apply_forward_scaling`-by-multiply-by-1.0 was coercing integer token-index inputs to float. The package layout, separately, had a true `nodes/base.py` ↔ `builder/__init__.py` cycle papered over by function-local imports. Both were getting harder to maintain as the node zoo grew (transformers, Hopfield, embeddings).

## What's in it

**Gradient path unified.** `NodeBase.forward_inference / forward_learning` → `forward_and_latent_grads / forward_and_weight_grads`. The inference variant now returns `(state, input_grads, self_grad)` — the self-latent gradient is computed via `jax.value_and_grad` over `(inputs, z_latent)` rather than accumulated inside `energy_functional`. New `fabricpc/core/scaling.py` and `fabricpc/core/learning.py` host the scaling primitives and Hebbian gradient computation; nodes themselves are scaling-unaware.

**muPC contract refined for non-scalable slots.** Edges into `is_variance_scalable=False` slots are omitted from the per-edge scaling dicts rather than getting `1.0`. Callsites treat missing keys as no-op pass-through, which preserves input dtype across the boundary — required for integer clamps into `EmbeddingNode`.

**Integer clamps end-to-end.** State initializers propagate the clamp dtype to `z_latent` on clamped nodes only; other `NodeState` fields stay float so the inference carry is type-stable. `EmbeddingNode` and `train_autoregressive` stop casting indices to float.

**Package restructure.** `builder/` is gone. `Edge`, `SlotRef`, `GraphNamespace` moved to `core/topology.py`; `graph()`, `TaskMap` to `graph_assembly/`; `update_node_in_state`, `set_latents_to_clamps` to `core/state_ops.py`; the `graph` package is renamed to `graph_initialization` (with `params_initializer.py` and `state_initializer.py`). The dependency DAG is now strictly upward: `core → nodes → graph_assembly → graph_initialization, training`.

**Cleanups.** Removed `compute_local_weight_gradients_ar` (duplicate), `GraphStructure._topological_sort` (duplicate), the empty `graph_net.py` shim, and forced-`float32` dtype handling in state init. `LinearExplicitGrad` extracted to its own module as the canonical example of a hand-rolled-gradient node under the new contract; `ActivationBase.jacobian()` added to support that pattern, with `SoftmaxActivation.jacobian()` implemented.

## Risk and verification
No mathematical changes to forward, energy, or scaling formulas — this is a refactor of where the computation lives, plus the dtype/identity-passthrough fix on non-scalable slots. Verified via the full test suite (127 passing) and numerical results match in the MNIST + transformer + resent + hopfield demos. Downstream forks: see the migration guide in `CHANGELOG.md` under `[0.3.1]` — the changes are mostly mechanical import rewrites and two node-method renames with one signature change (third return value).

+++ examples/transformer_v2_demo.py
@@ -29,7 +29,7 @@ set_jax_flags_before_importing_jax()
 import argparse
 import jax
 import jax.numpy as jnp
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.training import (
     train_pcn,
     evaluate_transformer,
@@ -37,7 +37,7 @@ from fabricpc.training import (
     evaluate_backprop,
 )
 from fabricpc.core.inference import run_inference, InferenceSGD
-from fabricpc.graph.state_initializer import initialize_graph_state
+from fabricpc.graph_initialization.state_initializer import initialize_graph_state
 from fabricpc.nodes.transformer_v2 import create_deep_transformer
 from fabricpc.utils.data import CharDataLoader
 import optax
@@ -114,7 +114,7 @@ def generate(
 
     print(f"--- Generating ---")
     for _ in range(length):
-        input_batch = jnp.array([current_indices], dtype=jnp.float32)
+        input_batch = jnp.array([current_indices], dtype=jnp.int32)
         inputs = {"input_ids": input_batch}
         batch_size = input_batch.shape[0]
